// src/api/generated/model/outreachChannel.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type OutreachChannel = typeof OutreachChannel[keyof typeof OutreachChannel];


export const OutreachChannel = {
  email: 'email',
  linkedin_inmail: 'linkedin_inmail',
  call_script: 'call_script',
} as const;