// src/api/generated/model/outreachTone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export type OutreachTone = typeof OutreachTone[keyof typeof OutreachTone];


export const OutreachTone = {
  professional: 'professional',
  conversational: 'conversational',
  direct: 'direct',
} as const;